move pid header and entry to seperate modules
